+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - fix: pas maxBreadcrumbs to Android init
 - feat: Allow disabling native SDK initialization but still use it #1259
+- ref: Rename shouldInitializeNativeSdk to autoInitializeNativeSdk #1275
 - fix: Fix parseErrorStack that only takes string in DebugSymbolicator event processor #1274
 - fix: Only set "event" type in envelope item and not the payload #1271
-
 
 ## 2.1.0